let's support zsh for real :)
`~/.profile` is not used for zsh FYI.

Instead, `~/.zprofile` is used, but this file is only for login shell. We need
to support interactive shell as well (I use PuTTY a lot...)

I understand that there's always a workaround for zsh users to source the
~/.profile, but that requires an extra step for users. Using this code would
make zsh users' life easier. :D